requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
datasets==1.18.4
dill==0.3.7
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==6.7.0
multidict==6.0.5
multiprocess==0.70.15
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pip==22.3.1
pyarrow==12.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
responses==0.18.0
scipy==1.7.3
setuptools==65.6.3
six==1.17.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.67.1
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.14
wheel==0.42.0
xxhash==3.6.0
yarl==1.9.4
zipp==3.15.0